Refactor Progress viz to TypeScript and a function component
